+++ go.mod
@@ -1,3 +1,3 @@
-module bank
+module github.com/ajtfj/bank
 
 go 1.18
